requirements.txt
tweepy==4.10.0
python-dotenv
pysqlite3